requirements.txt
feedparser
readability-lxml
html2text
pymongo[srv]
nltk
numpy
